feat: replace OrderBook with App.vue, remove OrderBook component, and add unit tests in App.test.js
