+++ vite.config.js
@@ -2,7 +2,7 @@ import { defineConfig } from 'vite';
 import vue from '@vitejs/plugin-vue';
 
 export default defineConfig({
-  base: '/vue-trade/',
+  base: '/',
   plugins: [vue()],
   resolve: {
     alias: {
